test-requirements.txt
ansible-core
black
coverage==4.5.4
flake8
git+https://github.com/ansible-network/pytest-ansible-network-integration.git
git+https://github.com/ansible-community/pytest-ansible-units.git
pre-commit
pytest-xdist
